refactor(cmd/reverst): move Server into own internal package
